refactor(link): implement stories
