+++ src/components/link/block-link-static/block-link-static.stories.ts
@@ -4,15 +4,23 @@ import { styleMap } from 'lit/directives/style-map.js';
 import { html } from 'lit/static-html.js';
 
 import {
-  inline,
-  inlineNegative,
+  blockFixedWidth,
+  blockIconStart,
+  blockM,
+  blockMIcon,
+  blockNegative,
+  blockS,
+  blockSIcon,
+  blockWithSlottedIcon,
+  blockXS,
+  blockXSIcon,
   linkCommonDefaultArgs,
   linkCommonDefaultArgTypes,
   wrapperStyle,
 } from '../common/link-common-stories';
 
 import readme from './readme.md?raw';
-import './link-static';
+import './block-link-static';
 
 const defaultArgTypes: ArgTypes = {
   ...linkCommonDefaultArgTypes,
@@ -20,11 +28,19 @@ const defaultArgTypes: ArgTypes = {
 
 const defaultArgs: Args = {
   ...linkCommonDefaultArgs,
-  tag: 'sbb-link-static',
+  tag: 'sbb-block-link-static',
 };
 
-export const Inline: StoryObj = inline;
-export const InlineNegative: StoryObj = inlineNegative;
+export const BlockXS: StoryObj = blockXS;
+export const BlockS: StoryObj = blockS;
+export const BlockM: StoryObj = blockM;
+export const BlockXSIcon: StoryObj = blockXSIcon;
+export const BlockSIcon: StoryObj = blockSIcon;
+export const BlockMIcon: StoryObj = blockMIcon;
+export const BlockIconStart: StoryObj = blockIconStart;
+export const BlockNegative: StoryObj = blockNegative;
+export const BlockWithSlottedIcon: StoryObj = blockWithSlottedIcon;
+export const BlockFixedWidth: StoryObj = blockFixedWidth;
 
 const meta: Meta = {
   argTypes: defaultArgTypes,
